+++ src/main/java/org/ws/parsepdf/ServiceRescheduleMoreTime.java
@@ -43,31 +43,36 @@ public class ServiceRescheduleMoreTime {
                                @FormDataParam("reschedule") boolean reschedule, @FormDataParam("password") String passwd, @FormDataParam("iter") int iter) {
       // System.out.println( p.getValue());
         if(resource != null && currentTime > -1 && "12345".equals(passwd) && iter > 0) {
-            readSolutionFile(iter,reschedule);//read solution file, constraints and time recordings
-            int job, op;
-            int machine = ServiceRescheduleLessResources.getMachine(resource);
-            int[] job_op = searchJobOP(machine, currentTime);//buscar que operation is
-            job = job_op[0];
-            op = job_op[1];
 
-            //asignar a timeReSchedule el tiempo a empezar la resecuenciacion
-            ql.Machines[machine].timeReSchedule = ql.Jobs[job].operations.get(op).initial_time + timeOperation;
-           // System.out.println("Job "+job+" op "+op+" Time ReSchedule "+ ql.Machines[machine].timeReSchedule);
-            Operation operation = ql.Jobs[job].operations.get(op);
-            startTimes(operation,timeOperation);
+            boolean solution = readSolutionFile(iter, reschedule);//read solution file, constraints and time recordings
+            if (solution) {
+                int job, op;
+                int machine = ServiceRescheduleLessResources.getMachine(resource);
+                int[] job_op = searchJobOP(machine, currentTime);//buscar que operation is
+                job = job_op[0];
+                op = job_op[1];
 
-            //New times of the operation
-            operation.end_time = operation.initial_time + timeOperation; //new operation's time initial
-            operation.proc_time = timeOperation;
-            try {
-                ql.ExecuteReSchedule();
-            } catch (FileNotFoundException | CloneNotSupportedException e) {
-                e.printStackTrace();
+                //asignar a timeReSchedule el tiempo a empezar la resecuenciacion
+                ql.Machines[machine].timeReSchedule = ql.Jobs[job].operations.get(op).initial_time + timeOperation;
+                // System.out.println("Job "+job+" op "+op+" Time ReSchedule "+ ql.Machines[machine].timeReSchedule);
+                Operation operation = ql.Jobs[job].operations.get(op);
+                startTimes(operation, timeOperation);
+
+                //New times of the operation
+                operation.end_time = operation.initial_time + timeOperation; //new operation's time initial
+                operation.proc_time = timeOperation;
+                try {
+                    ql.ExecuteReSchedule();
+                } catch (FileNotFoundException | CloneNotSupportedException e) {
+                    e.printStackTrace();
+                }
+                File solution1 = new File(String.format("%s/schedule.txt", UPLOAD_FOLDER));
+                return Response.status(200).entity(solution1).build();
+            } else {
+                return Response.status(400).entity("Invalid form data").build();
             }
-            return Response.status(200).entity("OK").build();
-        }else{
-            return Response.status(400).entity("Invalid form data").build();
-        }
+        }else
+            return Response.status(400).entity("Files don't exist").build();
        // return Response.status(200).entity("OK").build();
 
     }
@@ -89,20 +94,25 @@ public class ServiceRescheduleMoreTime {
     }
 
     //leer el archivo solution para saber dónde es la interrupcion
-    public void readSolutionFile(int iter, boolean reschedule) {
-        File solution = reschedule ? new File(String.format("%s/solution_copy.tmp", UPLOAD_FOLDER))
+    public boolean readSolutionFile(int iter, boolean reschedule){
+        File solution = null;
+        File[] targetFile = new File[2]; //files constraints and times recording
+        try {
+             solution = reschedule ? new File(String.format("%s/solution_copy.tmp", UPLOAD_FOLDER))
                 : new File(String.format("%s/solution.tmp", UPLOAD_FOLDER)); //Read before solution
 
-        File[] targetFile = new File[2]; //files constraints and times recording
         targetFile[0] = new File(String.format("%s/targetFile.txt", UPLOAD_FOLDER));
         targetFile[1] = new File(String.format("%s/targetFile2.txt", UPLOAD_FOLDER));
-        try {
-            if (!reschedule) {
+
+            if (!reschedule && solution.exists()) {
                 File solutionCopy = new File(String.format("%s/solution_copy.tmp", UPLOAD_FOLDER)); //crear copia de la solution para reschedule
                 FileUtils.copyFile(solution,solutionCopy); //copy solution to solution_copy
             }
 
-        } catch (Exception e) {
+        } catch (FileNotFoundException e) {
+            Response.status(500).entity("No existen ficheros soluciones en el server").build();
+            e.printStackTrace();
+        } catch (IOException e) {
             e.printStackTrace();
         }
 
@@ -111,13 +121,20 @@ public class ServiceRescheduleMoreTime {
         double epsi = 0.2;
 
         try {
-            ql = new QLearning(targetFile, LR, DF, iter, epsi);
-            ql.ReadData(targetFile);
-            new Test(solution,ql);
+            if (solution.exists()){
+                ql = new QLearning(targetFile, LR, DF, iter, epsi);
+                ql.ReadData(targetFile);
+                new Test(solution,ql);
+            }
            // ql.PrintJob();
+        } catch (FileNotFoundException e) {
+            Response.status(500).entity("No existen ficheros soluciones en el server").build();
+            e.printStackTrace();
         } catch (IOException e) {
+            Response.status(500).entity("Ficheros incorrectos").build();
             e.printStackTrace();
         }
+        return solution.exists();
     }
 
     private void startTimes(Operation operationFix, int newTime) {
